Full Card Data Revamp
Basic concept of feature: Now uses a data structure to use with a data table that keeps track of all the cards in the game. The data structure has two variables which define the type of the cards and its card index number to create a unique ID when looking for cards. Also uses an enum to perform actions based on the type of card. More of this will be explained in documentation.