Character IDs
Added a way to assign character information for players and AI to avoid the need to rely on things such as "get player pawn [0]" which is bad practice. This allows pawns to have an owning character which can then be used to make various checks. This commit does not include code for when a hero dies, but will be done later.